chore(base): Move everything related to `Room` members in `members`.
This patch moves all types and methods used by or implemented on `Room`
inside the existing `members` module.
